gospace: Uniform error signature for interfaces.
External and internal interfaces that deal with spaces now have an
almost uniform method signature. There is an internal error model which
will be used for debugging internal interfaces, for external interfaces
implementing "error" interface and providing useful error messages.

The return values can of course be ignore, but they return a value and
an error. If an error does not occur, a structure is returned which can
be used to print out an error message by using, say, fmt.Printf.
Otherwise, if no error occurs, nil is returned.

If there will be a need for more powerful inspection of the errors,
an extended error interface will be provided.

gospace.go: Updated interface documentation.

{template, tuple}.go: Small refactoring and documentation update.

space.go:
    - Added documentation for all methods operating on a space.
    - Moved internal implementation of NewSpace and NewRemoteSpace to
    tuplespace_alt.go.
    - Ractored interfaces.
    - Introducing the Interstellar interface:
        * This is an internal interface which can change and should not
        be considered stable.
        * Its primary goal is to define raw operations which can be
          performed on spaces without any error checking. This will be
          used to test operation performance on spaces and to see if error
          checking has heavy impact on usage.
        * It also exposes the implementation result and errors via
          interface{} which can be handy in debugging scenarios,
          and hacking on the implementation to test new ideas out.
    - Added error functions InterpretError() and InterpretOperation()
      for the internal interfaces.
        * These functions are used by the error mechanisms in
          space_error.go.
        * It provides a convinient way of interpreting error state returned
          by operations, without adding repeatedly the same generic error
          checking code everywhere.
    - Added utility function TupleFromTemplate().

space_error.go: An error model for the internal interfaces.

This defines error types (and in the future possible error interfaces), for
debugging internal interfaces such as Interspace and Interstellar.

Its goal is to give meaningful errors back and interpret the returned
state by the various operations. It somewhat generic, meaning that it
could be extended to other structures, but for now, spaces are
privileged to have its own error structure.

The error types implement the "error" interface and can be used by
external interfaces, but there is no guarantees for stability of the
error types.

tuplespace_alt.go: Brought back the constructs which are purely
implementation related.