gospace: Added proper Space URI parsing.
pointtopoint.go: Changed method signature.

space_uri*.go: Space URI parsing according to pSpace specification.

This allows to be used for spaces and gates. If the URI is not sane,
nil is returned. If parsing of the URI, an error is retunred.

tuple.go: Changed TupleFromTemplate() to CreateTupleFromTemplate().

It is now possible to create tuples from templates.

space.go:
    - Changed method call signatures and the Space structure.
    - Moved TupleFromTemplate() to tuple.go.
    - Fixed errors where one would always be returing tuples no
      matter if there were errors or not.
    - InterpretValue() is an internal method used to interpret
      parameter values for a space operation.

space_error.go: Improved error messages with an value parameter.

The value for now the parameter list for an operation. This makes it
easier to spot errors when debugging and doing diagnostics.

README.md: Updated documentation.

tuplespace_alt.go:
    - Incorporated usage of Space URIs.
    - Initial fix for hanging of when using Query().
    - For now, gospace is limited to TCP over IPv4.
        * Because the protocols are not properly abstracted, multiple
          resolution of host names is not supported yet.